--- a/lib/assert/context/test_dsl.rb
+++ b/lib/assert/context/test_dsl.rb
@@ -15,23 +15,23 @@ class Assert::Context
         # create a test from the given code block
         self.suite.on_test(Assert::Test.for_block(
           desc_or_macro.kind_of?(Assert::Macro) ? desc_or_macro.name : desc_or_macro,
-          Assert::ContextInfo.new(self, called_from, first_caller || caller.first),
+          Assert::ContextInfo.new(self, called_from, first_caller || caller_locations.first),
           self.suite.config,
           &block
         ))
       else
-        test_eventually(desc_or_macro, called_from, first_caller || caller.first, &block)
+        test_eventually(desc_or_macro, called_from, first_caller || caller_locations.first, &block)
       end
     end
 
     def test_eventually(desc_or_macro, called_from = nil, first_caller = nil, &block)
       # create a test from a proc that just skips
-      ci = Assert::ContextInfo.new(self, called_from, first_caller || caller.first)
+      ci = Assert::ContextInfo.new(self, called_from, first_caller || caller_locations.first)
       self.suite.on_test(Assert::Test.for_block(
         desc_or_macro.kind_of?(Assert::Macro) ? desc_or_macro.name : desc_or_macro,
         ci,
         self.suite.config,
-        &proc { skip('TODO', ci.called_from) }
+        &proc { skip('TODO', [ci.called_from.to_s]) }
       ))
     end
     alias_method :test_skip, :test_eventually
@@ -40,14 +40,14 @@ class Assert::Context
       if !desc_or_macro.kind_of?(Assert::Macro)
         desc_or_macro = "should #{desc_or_macro}"
       end
-      test(desc_or_macro, called_from, first_caller || caller.first, &block)
+      test(desc_or_macro, called_from, first_caller || caller_locations.first, &block)
     end
 
     def should_eventually(desc_or_macro, called_from = nil, first_caller = nil, &block)
       if !desc_or_macro.kind_of?(Assert::Macro)
         desc_or_macro = "should #{desc_or_macro}"
       end
-      test_eventually(desc_or_macro, called_from, first_caller || caller.first, &block)
+      test_eventually(desc_or_macro, called_from, first_caller || caller_locations.first, &block)
     end
     alias_method :should_skip, :should_eventually