--- a/lib/assert/context_info.rb
+++ b/lib/assert/context_info.rb
@@ -7,7 +7,7 @@ module Assert
     def initialize(klass, called_from = nil, first_caller = nil)
       @called_from = called_from || first_caller
       @klass = klass
-      @file = @called_from.gsub(/\:[0-9]+.*$/, '') if @called_from
+      @file = @called_from.to_s.gsub(/\:[0-9]+.*$/, '') if @called_from
     end
 
     def test_name(name)