@kellyredding 2018-05-12
@jcredding so you are good with the polyfill approach - I didn't overlook any concerns there?