@jcredding 2018-05-12
@kellyredding - I think the polyfill is fine in this case. The only concern to me would be you are now defining a method on `Kernel` so its available in other libraries/apps. We could become dependent on it in another library/app and think everything is good because tests pass but then it errors outside of the test suite. I think this is very unlikely though because of the method; `caller` is not the most common method and I think we're unlikely to change anything currently using it.